+++ packages/core/src/vault/index.ts
@@ -1,3 +1,4 @@
+export * from "./get.js";
 export * from "./list.js";
 export * from "./upsert.js";
 export * from "./delete.js";